+++ HISTORY.rst
@@ -4,9 +4,10 @@ History
 3.11.1 (unreleased)
 -------------------
 
-- Add CSV upload support for 1D initial water level (#575)
-- Add initial concentrations to initial conditions page (#580)
-- Fixes/enhancements: #238, #391, #480, #493, #496, #526
+- Add CSV upload support for 1D initial water level (#575).
+- Add initial concentrations to initial conditions page (#580).
+- Convert timesteps to correct time units (#585).
+- Fixes/enhancements: #238, #391, #480, #493, #496, #526.
 
 
 3.11.0 (2024-06-21)